leaves_ograph/requirements.txt
Flask
gunicorn
newspaper3k
nltk
flask_cors